moves resolve tests
